Initial Upload of GROUPS
The GROUPS script will download and organize a list of every group in a course and which students aren't in a group for each group set. Additionally you can also ask it to check if each student in each group is in the same section. The output is a csv file.